backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
python-dotenv
httpx
openai
python-multipart
